users/views.py
from django.contrib.auth.models import User
from django.db import transaction
from rest_framework import generics
from rest_framework import permissions
from rest_framework.response import Response
from rest_framework import status

from users.serializer import AccountSerializer


class AuthRegister(generics.CreateAPIView):
    permission_classes = (permissions.AllowAny,)
    queryset = User.objects.all()
    serializer_class = AccountSerializer

    @transaction.atomic
    def post(self, request, format=None):
        serializer = AccountSerializer(data=request.data)

        if serializer.is_valid():
            serializer.save()
            return Response(serializer.data, status=status.HTTP_201_CREATED)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
