requirements.txt
autoflake==1.1
autopep8==1.3.5
Django==2.0.4
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docformatter==1.0
flake8==2.5.5
hacking==1.0.0
mccabe==0.2.1
pbr==4.0.2
pep8==1.5.7
pycodestyle==2.4.0
pyflakes==1.6.0
pyformat==0.7
PyJWT==1.6.1
pytz==2018.4
six==1.11.0
unify==0.4
untokenize==0.1.1
